Transcripcion/requirements.txt
openai-whisper>=20231117
sumy>=0.11.0
nltk>=3.8.1
streamlit>=1.36.0
fastapi>=0.111.0
uvicorn>=0.30.1
python-multipart>=0.0.9
